Warnings.

+++ src/hashes/rmd256.c
@@ -407,7 +407,7 @@ int rmd256_test(void)
 
    for (i = 0; i < (int)LTC_ARRAY_SIZE(tests); i++) {
        rmd256_init(&md);
-       rmd256_process(&md, (unsigned char *)tests[i].msg, XSTRLEN(tests[i].msg));
+       rmd256_process(&md, (unsigned char *)tests[i].msg, (unsigned long)XSTRLEN(tests[i].msg));
        rmd256_done(&md, tmp);
        if (ltc_compare_testvector(tmp, sizeof(tmp), tests[i].hash, sizeof(tests[i].hash), "RIPEMD256", i)) {
           return CRYPT_FAIL_TESTVECTOR;
